Move tests to unit test folder.
