+++ system-test/read-rows.ts
@@ -218,62 +218,4 @@ describe('Bigtable/Table', () => {
       });
     });
   });
-
-  describe('createReadStream mocking out the gapic layer', () => {
-    // TODO: Add true/false checker for the canResume function.
-    // TODO: Add checker for return value of the resumption function.
-
-    // TODO: Consider moving this to unit tests.
-    // TODO: Write tests to ensure options reaching Gapic layer are
-    // TODO: Future tests
-    // 1. Provide more gax options
-    // 2. Override the retry function
-    // 3. Anything with retryRequestOptions?
-    // unchanged for other streaming calls
-    const bigtable = new Bigtable({
-      projectId: 'fake-project-id',
-    });
-    const tester = new GapicLayerTester(bigtable);
-    const table: Table = bigtable.instance('fake-instance').table('fake-table');
-    const tableName =
-      'projects/fake-project-id/instances/fake-instance/tables/fake-table';
-
-    it('should pass the right retry configuration to the gapic layer', done => {
-      const expectedOptions = tester.buildReadRowsGaxOptions(tableName, {});
-      tester.testReadRowsGapicCall(
-        done,
-        {
-          rows: {
-            rowKeys: [],
-            rowRanges: [{}],
-          },
-          tableName,
-        },
-        expectedOptions
-      );
-      table.createReadStream();
-    });
-    it('should pass maxRetries to the gapic layer', done => {
-      const expectedOptions = Object.assign(
-        {maxRetries: 7},
-        tester.buildReadRowsGaxOptions(tableName, {})
-      );
-      tester.testReadRowsGapicCall(
-        done,
-        {
-          rows: {
-            rowKeys: [],
-            rowRanges: [{}],
-          },
-          tableName,
-        },
-        expectedOptions
-      );
-      const tableWithRetries: Table = bigtable
-        .instance('fake-instance')
-        .table('fake-table');
-      tableWithRetries.maxRetries = 7;
-      tableWithRetries.createReadStream();
-    });
-  });
 });
